testing_requirements.txt
pytest==5.0.1
pytest-env==0.6.2
pytest-cov==2.8.1
pytest-django==3.9.0
djangorestframework==3.11.0
